Refactored kit
